# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Developer News Dashboard API",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get Me",
        "description": "Current user and their preference profile.",
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete Me",
        "description": "Delete the account and cascade preferences, bookmarks, reads, and\nsessions; then sign out.",
        "operationId": "delete_me_api_v1_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/preferences": {
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Patch Preferences",
        "description": "Partial preference update in one transaction.",
        "operationId": "patch_preferences_api_v1_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Shape of FastAPI's HTTPException body; used in OpenAPI responses."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Layout": {
        "type": "string",
        "enum": [
          "grid",
          "list"
        ],
        "title": "Layout"
      },
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesOut"
          }
        },
        "type": "object",
        "required": [
          "user",
          "preferences"
        ],
        "title": "MeResponse"
      },
      "PreferencesOut": {
        "properties": {
          "theme": {
            "$ref": "#/components/schemas/Theme"
          },
          "layout": {
            "$ref": "#/components/schemas/Layout"
          },
          "max_visible_cards": {
            "type": "integer",
            "title": "Max Visible Cards"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics",
            "description": "Selected topic slugs"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "Enabled source slugs"
          }
        },
        "type": "object",
        "required": [
          "theme",
          "layout",
          "max_visible_cards",
          "onboarding_completed",
          "topics",
          "sources"
        ],
        "title": "PreferencesOut"
      },
      "PreferencesPatch": {
        "properties": {
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Theme"
              },
              {
                "type": "null"
              }
            ]
          },
          "layout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Layout"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_visible_cards": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Visible Cards"
          },
          "onboarding_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed"
          },
          "topics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topics"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "type": "object",
        "title": "PreferencesPatch",
        "description": "Partial update: absent fields are left untouched. An explicit\nempty list clears the corresponding selection."
      },
...
      "Theme": {
        "type": "string",
        "enum": [
          "light",
          "dark",
          "system"
        ],
        "title": "Theme"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_id": {
            "type": "integer",
            "title": "Github Id"
          },
          "github_login": {
            "type": "string",
            "title": "Github Login"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_id",
          "github_login",
          "display_name",
          "avatar_url",
          "is_admin",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}